throw error is invalid input is provied and refactor code
